requirements.txt
Django==3.2
django-qr-code==2.1.0
Pillow==8.2.0
xhtml2pdf==0.2.5
